refactor(tx-flow): remove createTxFlow and implement  less complex TxFlow component instead
