+++ apps/web/src/components/tx-flow/actions/Propose.tsx
@@ -1,7 +1,7 @@
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
 import { useCallback, useContext } from 'react'
 import { TxFlowContext } from '../TxFlowProvider'
-import type { SubmitCallback } from '../createTxFlow'
+import type { SubmitCallback } from '../TxFlow'
 import ProposerForm from '@/components/tx/SignOrExecuteForm/ProposerForm'
 
 type ProposeProps = {
